Update LO - Circle in Square.c

+++ LO - Circle in Square.c
@@ -2,15 +2,15 @@
 #include <math.h>
 main()
 {
-    float T,circle_area,blue_part,square_area,pi=2*acos(0),r;
-    scanf("%f",&T);
+    double T,circle_area,blue_part,square_area,pi=2*acos(0),r;
+    scanf("%lf",&T);
     for(int i=1;i<=T;i++)
     {
-        scanf("%f",&r);
+        scanf("%lf",&r);
         circle_area=pi*r*r;
         square_area=(r*2)*(r*2);
         blue_part=square_area-circle_area;
-        printf("Case %d: %.2f\n",i,blue_part);
+        printf("Case %d: %.2lf\n",i,blue_part);
     }
     return 0;
 }
